Adicionar organizacao da extensao
