+++ macroChallengeApp/macroChallengeAppTests/macroChallengeAppTests.swift
@@ -9,7 +9,6 @@ import XCTest
 @testable import macroChallengeApp
 
 class MacroChallengeAppTests: XCTestCase {
-
     override func setUpWithError() throws {
         // Put setup code here. This method is called before the invocation of each test method in the class.
     }
