+++ macroChallengeApp/macroChallengeAppUITests/macroChallengeAppUITestsLaunchTests.swift
@@ -8,7 +8,6 @@
 import XCTest
 
 class MacroChallengeAppUITestsLaunchTests: XCTestCase {
-
     override class var runsForEachTargetApplicationUIConfiguration: Bool {
         true
     }
